Add files via upload
cluster_startL.ipynb is the code I originally started to manipulate which include code that generalizes the 3 most common colors of an image. I faced difficulties with this code working. The silhouette score would not change regardless of the data features I chose. 

cluster_start.ipynb is the restarted and final version.